Fix bug in seedDao findBySeedNameLike method
The findBySeedNameLike method in seedDao had a bug related to fuzzy
searching the seedName in the database. To resolve this issue, the
method required manually adding '%' at both ends of the username.
However, I have made the following changes to improve this
functionality:

- Replaced the method with findBySeedNameContaining, eliminating the
need to manually add '%'.
- Updated the method name to findBySeedNameContaining for accurate
seedName searching.

This commit addresses the bug and enhances the seedName search
capability in the database.