Bug Fix: Resolved issue with missing 24 farm lands in landStatus
This commit fixes a bug in the "/game/landStatus" endpoint that
caused an empty list to be returned when a user had not planted
any crops or was a new user. This resulted in difficulties for
the frontend to handle the response.

To address this issue, a check has been added after retrieving
data from the database to ensure that the list contains all 24
farm lands. If the list is incomplete, empty farm lands are
inserted into the database to complete the set of 24.

- Added check to ensure the presence of all 24 farm lands in
  the landStatus response
- Inserted empty farm lands into the database if needed